requirements.txt
# Flask Web Framework
Flask>=2.3.0
Werkzeug>=2.3.0

# Computer Vision and Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# YOLOv8 Object Detection
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# OCR (Optical Character Recognition)
pytesseract>=0.3.10
easyocr>=1.7.0

# Security
bcrypt>=4.0.0

# Development and Testing
pytest>=7.4.0
pytest-flask>=1.2.0

# Optional: For better performance in production
gunicorn>=21.2.0

# Optional: For environment variable management
python-dotenv>=1.0.0
